Parse operations for mempool transactions
