+++ core-rust/mesh-api-server/src/mesh_api/handlers/construction_parse.rs
@@ -1,8 +1,4 @@
 use crate::prelude::*;
-use radix_engine_interface::blueprints::account::{
-    AccountTryDepositOrAbortManifestInput, AccountWithdrawManifestInput,
-};
-use radix_transactions::manifest::{CallMethod, TakeFromWorktop};
 use radix_transactions::validation::TransactionValidator;
 
 // This method only accepts transactions constructed with the Mesh API,
@@ -53,7 +49,7 @@ pub(crate) async fn handle_construction_parse(
 
     let mapping_context = MappingContext::new(&state.network);
     let database = state.state_manager.database.snapshot();
-    let operations = parse_instructions(&instructions, &mapping_context, database.deref())?;
+    let operations = to_mesh_api_operations_from_instructions_v1(&instructions, &mapping_context, database.deref())?;
 
     // See https://docs.cdp.coinbase.com/mesh/docs/models#constructionparseresponse for field
     // definitions
@@ -70,98 +66,4 @@ pub(crate) async fn handle_construction_parse(
         ),
         metadata: None,
     }))
-}
-
-pub fn parse_instructions(
-    instructions: &[InstructionV1],
-    mapping_context: &MappingContext,
-    database: &StateManagerDatabase<impl ReadableRocks>,
-) -> Result<Vec<models::Operation>, ResponseError> {
-    let mut operations = Vec::new();
-    let mut next_index = 0;
-    while next_index < instructions.len() {
-        let mut instruction = &instructions[next_index];
-        next_index = next_index + 1;
-        match instruction {
-            InstructionV1::CallMethod(CallMethod {
-                address: DynamicGlobalAddress::Static(global_address),
-                method_name,
-                args,
-            }) if global_address.is_account() => {
-                let args_bytes = manifest_encode(&args).unwrap();
-                match method_name.as_str() {
-                    "lock_fee" => (),
-                    "withdraw" => {
-                        let input = manifest_decode::<AccountWithdrawManifestInput>(&args_bytes)
-                            .map_err(|_| {
-                                ResponseError::from(ApiError::InvalidWithdrawInstruction)
-                                    .with_details("Invalid withdraw instruction")
-                            })?;
-                        operations.push(to_mesh_api_operation_no_fee(
-                            mapping_context,
-                            database,
-                            operations.len() as i64,
-                            None,
-                            global_address,
-                            &match input.resource_address {
-                                ManifestResourceAddress::Static(resource_address) => {
-                                    resource_address
-                                }
-                                ManifestResourceAddress::Named(_) => {
-                                    return Err(ResponseError::from(
-                                        ApiError::NamedAddressNotSupported,
-                                    )
-                                    .with_details("Named address is not supported"))
-                                }
-                            },
-                            -input.amount.clone(),
-                        )?);
-                    }
-                    _ => {
-                        return Err(ResponseError::from(ApiError::UnrecognizedInstruction)
-                            .with_details(format!("Unrecognized instruction: {:?}", instruction)));
-                    }
-                }
-            }
-            InstructionV1::TakeFromWorktop(TakeFromWorktop {
-                resource_address,
-                amount,
-            }) if next_index < instructions.len() => {
-                instruction = &instructions[next_index];
-                next_index = next_index + 1;
-
-                match instruction {
-                    InstructionV1::CallMethod(CallMethod {
-                        address: DynamicGlobalAddress::Static(global_address),
-                        method_name,
-                        args,
-                    }) if method_name.eq("try_deposit_or_abort") && global_address.is_account() => {
-                        if let Ok(_input) = manifest_decode::<AccountTryDepositOrAbortManifestInput>(
-                            &manifest_encode(&args).unwrap(),
-                        ) {
-                            operations.push(to_mesh_api_operation_no_fee(
-                                mapping_context,
-                                database,
-                                operations.len() as i64,
-                                None,
-                                global_address,
-                                resource_address,
-                                *amount,
-                            )?);
-                        }
-                    }
-                    _ => {
-                        return Err(ResponseError::from(ApiError::UnrecognizedInstruction)
-                            .with_details(format!("Unrecognized instruction: {:?}", instruction)));
-                    }
-                }
-            }
-            _ => {
-                return Err(ResponseError::from(ApiError::UnrecognizedInstruction)
-                    .with_details(format!("Unrecognized instruction: {:?}", instruction)));
-            }
-        }
-    }
-
-    Ok(operations)
 }
